Add toner sales history to client view
Introduces a new section in the client page displaying a complete history of toners sold to the client. Updates backend to provide aggregated toner sales data and frontend to render this information in a grid of blocks. Also enhances purchase item display to show quantity and sale value per toner.